+++ src/registers.rs
@@ -1,5 +1,7 @@
 pub use sx127x_common::registers::*;
 
+// TODO registers pub, masks pub(crate)
+
 // RegOpMode ---------------------------------------------------------------------------------------
 pub const OP_MODE_LONG_RANGE_MODE_MASK: u8 = 0x80;
 pub const OP_MODE_MODULATION_TYPE_MASK: u8 = 0x60;
@@ -24,8 +26,12 @@ pub const RX_CONFIG_RX_TRIGGER_MASK: u8 = 0x07;
 
 // RegRssiConfig -------------------------------------------------------------------------------------
 pub const RSSI_CONFIG: u8 = 0x0e;
-pub const RSSI_CONFIG_RSSI_OFFSET_MASK: u8 = 0xf8; // TODO impl
+pub const RSSI_CONFIG_RSSI_OFFSET_MASK: u8 = 0xf8;
 pub const RSSI_CONFIG_RSSI_SMOOTHING_MASK: u8 = 0x07;
+
+// -------------------------------------------------------------------------------------------------
+pub const RSSI_COLLISION: u8 = 0x0f;
+pub const RSSI_THRESH: u8 = 0x10;
 
 // RegRxBw -----------------------------------------------------------------------------------------
 pub const RX_BW: u8 = 0x12;